Move device widget inside collections/device
