+++ lib/ui/collections/device/device_folder_icon_widget.dart
