+++ lib/ui/collections/device/device_folders_grid_view_widget.dart
@@ -9,7 +9,7 @@ import 'package:photos/events/backup_folders_updated_event.dart';
 import 'package:photos/events/local_photos_updated_event.dart';
 import "package:photos/generated/l10n.dart";
 import 'package:photos/models/device_collection.dart';
-import 'package:photos/ui/collections/device_folder_icon_widget.dart';
+import 'package:photos/ui/collections/device/device_folder_icon_widget.dart';
 import 'package:photos/ui/common/loading_widget.dart';
 import 'package:photos/ui/viewer/gallery/empty_state.dart';
