+++ lib/ui/tabs/user_collections_tab.dart
@@ -20,7 +20,7 @@ import "package:photos/ui/collections/button/hidden_button.dart";
 import "package:photos/ui/collections/button/trash_button.dart";
 import "package:photos/ui/collections/button/uncategorized_button.dart";
 import "package:photos/ui/collections/create_new_album_widget.dart";
-import 'package:photos/ui/collections/device_folders_grid_view_widget.dart';
+import 'package:photos/ui/collections/device/device_folders_grid_view_widget.dart';
 import 'package:photos/ui/collections/horizontal_grid_view.dart';
 import "package:photos/ui/collections/vertical_grid_view.dart";
 import 'package:photos/ui/common/loading_widget.dart';
